+++ 1.0/io/io/write_fits.h
@@ -16,15 +16,15 @@
 namespace io {
 
 /**
- * writes a double array into an image FITS file
+ * writes a float array into an image FITS file
  * @param[in] pathname name of the image FITS file to be written
- * @param[in] array double array to be written into the FITS file
+ * @param[in] array float array to be written into the FITS file
  * @param[in] name describes the content of the image
  * @param[in] naxis number of axes in the image
  * @param[in] naxes array of the axes of the image
  * @return WRITE_OK if all went well; an error message else
  */
-std::string write_image(std::string pathname, double *array, std::string name, int naxis, long *naxes);
+std::string write_image(std::string pathname, float *array, std::string name, int naxis, long *naxes);
 
 /**
  *
@@ -36,7 +36,7 @@ std::string write_image(std::string pathname, double *array, std::string name, i
  * @param naxes_2D
  * @return
  */
-std::string write_image_3D(std::string pathname_3D, double *array_2D_real, double *array_2D_imag, std::string name_3D, int naxis_2D, long *naxes_2D);
+std::string write_image_3D(std::string pathname_3D, float *array_2D_real, float *array_2D_imag, std::string name_3D, int naxis_2D, long *naxes_2D);
 
 } /* namespace io */
